requirements.txt
autopep8==1.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
idna==3.3
joblib==1.1.0
nltk==3.7
oauthlib==3.2.0
pycodestyle==2.8.0
python-dotenv==0.20.0
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
termcolor==1.1.0
textblob==0.17.1
toml==0.10.2
tqdm==4.64.0
urllib3==1.26.9
googletrans==4.0.0-rc1